Limit the maximum recv buffer size in read()
`rbufsize` is now used as the maximum size to with recv instead of
`left` because pyOpenSSL's recv typically reads just a small amount of
data (1KB observed) and returns immediately.

Even though it only returns a small amount of data each call, recv
still allocates a buffer of the requested size, which was causing
signficant performance problems when `size` - and therefore `left` -
is large. recv is called many times and each call was making time
consuming allocations of large buffers.

Using a small size with the recv calls doesn't change the number of
calls to recv but ensures each call is not making unnecessarily large
memory allocations.